Add files via upload
This dataset belongs to a banking domain where there are many attributes are given about the person borrowing money from the bank. and bank manager is highly interested to know whether the person will be a good candidate to lend money too or will he turn into a defaulter causing losses to the bank. Hence these candidates should be rejected for loan application. The intent of the model is to help manager to maximize the impact of his decision.